feat!: impersonate users (#3042)
* feat: added contextId to user data classes

* feat: contextId in H5PPlayer.play

* feat: contextId in h5p-express and docs

* test: added integration tests for contextId

* test: contextId in rest example

* feat(mongo-s3): added contextId to user data storage

* refactor: permission system

* refactor: permissions system

* test: fix tests

* test: fix more tests

* test: fix more tests

* refactor!: moved permissions from IUser to IPermissionSystem

* refactor: permission system

* refactor: fine grained permission types and REST example

* refactor: prettier

* refactor: more fine grained user data permission handling

* feat: impersonate users and read only states

* test: fixed test for new content user data interface

* refactor: style improvement and docs

* feat: add locales

* test: fix tests

* docs: extended docs

* test: added test for impersonation

* test: added impersonation and read only state to REST example

+++ packages/h5p-server/src/implementation/fs/DirectoryTemporaryFileStorage.ts
@@ -58,14 +58,19 @@ export default class DirectoryTemporaryFileStorage
 
     private maxFileLength: number;
 
-    public async deleteFile(filename: string, userId: string): Promise<void> {
+    public async deleteFile(filename: string, ownerId: string): Promise<void> {
         checkFilename(filename);
-        checkFilename(userId);
-        const filePath = this.getAbsoluteFilePath(userId, filename);
+        if (!ownerId) {
+            throw new Error(
+                'Invalid arguments for DirectoryTemporaryFileStorage.deleteFile: you must specify an ownerId'
+            );
+        }
+        checkFilename(ownerId);
+        const filePath = this.getAbsoluteFilePath(ownerId, filename);
         await fsExtra.remove(filePath);
         await fsExtra.remove(`${filePath}.metadata`);
 
-        const userDirectoryPath = this.getAbsoluteUserDirectoryPath(userId);
+        const userDirectoryPath = this.getAbsoluteUserDirectoryPath(ownerId);
         const fileDirectoryPath = path.dirname(filePath);
         if (userDirectoryPath !== fileDirectoryPath) {
             await this.deleteEmptyDirectory(fileDirectoryPath);
